+++ bugic/crates/bugic-core/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "bugic-core"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+thiserror.workspace = true
+bugic-share.workspace = true
