+++ bugic/crates/bugic-host/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "bugic-host"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+bugic-core.workspace = true
