+++ bugic/crates/bugic-host/src/lib.rs
@@ -1,17 +1,16 @@
 use std::collections::HashMap;
 
-use bugic_share::{ParamListFrom, SerializeTag, ToByte, ERROR_ABI_ID};
-
-use crate::{
-    plugin::PluginSystem,
-    BugiError, PluginSymbol,
+use bugic_core::{ParamListFrom, SerializeTag, ToByte, ERROR_ABI_ID};
+use bugic_core::{
+    PluginSystem,
+    BugiError,
 };
 
 pub(crate) type HostPluginFuncRaw = Box<dyn (Fn(&[u8]) -> Vec<u8>) + Send + Sync>;
 
 #[derive(Default)]
 pub struct HostPlugin {
-    funcs: HashMap<PluginSymbol, (u8, u8, HostPluginFuncRaw)>,
+    funcs: HashMap<String, (u8, u8, HostPluginFuncRaw)>,
 }
 
 impl HostPlugin {
@@ -27,11 +26,11 @@ impl HostPlugin {
         F: Fn(Param) -> Result + 'static + Send + Sync,
     >(
         &mut self,
-        symbol: String,
+        symbol: &str,
         func: F,
     ) {
         self.funcs.insert(
-            symbol,
+            symbol.to_string(),
             (
                 SInput::get_abi_id(),
                 SOutput::get_abi_id(),
@@ -46,7 +45,7 @@ impl HostPlugin {
 }
 
 impl PluginSystem for HostPlugin {
-    fn raw_call(&self, symbol: &PluginSymbol, param: &[u8], abi_arg: u8, abi_res: u8) -> Result<Vec<u8>, BugiError> {
+    fn raw_call(&self, symbol: &str, param: &[u8], abi_arg: u8, abi_res: u8) -> Result<Vec<u8>, BugiError> {
         let func = self.funcs.get(symbol).ok_or(BugiError::PluginCallError("Symbol not found".to_string()))?;
 
         if abi_arg != func.0 || abi_res != func.1 {
@@ -58,7 +57,7 @@ impl PluginSystem for HostPlugin {
 
     fn check_symbol_abi(
         &self,
-        symbol: &PluginSymbol,
+        symbol: &str,
         abi_arg: u8,
         abi_res: u8,
     ) -> Result<(), (u8, u8)> {
@@ -70,3 +69,4 @@ impl PluginSystem for HostPlugin {
         }
     }
 }
+
