+++ bugic/crates/bugic/src/plugin.rs
@@ -1,11 +1,7 @@
 use std::sync::Weak;
 
+use bugic_core::{BugiError, PluginSystem};
 use bugic_share::{FromByte, ParamListTo, SerializeTag};
-
-use crate::{
-    host_plug::HostPlugin,
-    BugiError, PluginSymbol,
-};
 
 /// plugin (original)
 pub struct Plugin {
@@ -15,36 +11,14 @@ pub struct Plugin {
 
 impl Plugin {
     /// Create a new Host Plugin
-    pub fn make_host(str_id: String, host: HostPlugin) -> Self {
-        Self {
-            str_id,
-            detail: Box::new(host),
-        }
+    pub fn make_plugin(str_id: &str, detail: impl PluginSystem + 'static) -> Self {
+        Self { str_id: str_id.to_string(), detail: Box::new(detail) }
     }
 
     /// Get the string ID of the plugin
     pub fn get_str_id(&self) -> &String {
         &self.str_id
     }
-}
-
-pub trait PluginSystem: Send + Sync {
-    /// call a plugin function
-    fn raw_call(
-        &self,
-        symbol: &PluginSymbol,
-        param: &[u8],
-        abi_arg: u8,
-        abi_res: u8,
-    ) -> Result<Vec<u8>, BugiError>;
-
-    /// check the ABI of a symbol
-    fn check_symbol_abi(
-        &self,
-        symbol: &PluginSymbol,
-        abi_arg: u8,
-        abi_res: u8,
-    ) -> Result<(), (u8, u8)>;
 }
 
 /// Reference to a plugin
@@ -67,19 +41,19 @@ impl PluginRef {
         Output: FromByte<SOutput>,
     >(
         &self,
-        symbol: String,
+        symbol: &str,
         param: Param,
     ) -> Result<Output, BugiError> {
         let plug = self.pref.upgrade().ok_or(BugiError::PluginDropped)?;
 
         plug.detail
-            .check_symbol_abi(&symbol, SInput::get_abi_id(), SOutput::get_abi_id())
+            .check_symbol_abi(symbol, SInput::get_abi_id(), SOutput::get_abi_id())
             .map_err(|(inp, out)| BugiError::PluginAbiError(inp, out))?;
 
         let param = param.to_byte().map_err(BugiError::CannotSerialize)?;
         let result =
             plug.detail
-                .raw_call(&symbol, &param, SInput::get_abi_id(), SOutput::get_abi_id())?;
+                .raw_call(symbol, &param, SInput::get_abi_id(), SOutput::get_abi_id())?;
         Ok(Output::from_byte(&result)?)
     }
 }
